pyproject.toml
[tool.poetry]
name = "mango-api"
version = "0.1.0"
description = ""
authors = ["Labidahrom <labidahrom@yandex.ru>"]
readme = "README.md"
packages = [{include = "mango_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pytz = "^2023.3.post1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
